Implement PA10-006y (#878): per-symbol @align(N) attribute for let declarations
Add `@align(N)` postfix syntax to item-level `let` declarations for
explicit symbol alignment override. Drives CR3-compatible page-table
allocation: `let mut pml4 : [u64; 512] = uninit @align(4096)` now
emits a 4096-byte-aligned `.bss` symbol.

## Grammar

    LetItem   ::= 'pub'? 'let' 'mut'? Name GenericParams? (':' Type)? '=' Expr AlignAttr? ';'?
    AlignAttr ::= '@' 'align' '(' IntLit ')'

Supported N range: any power of two in [1, 2^30]. Typical values: 8
(u64), 16 (SSE), 64 (cache line), 4096 (page), 2^21 (huge page),
2^30 (1 GiB).

## Threading through the toolchain

1. Parser (parse_item.rs): parse_optional_align_suffix helper after
   value parsing, before semicolon. Emits diagnostics P0250/P0251/P0252
   on unknown attribute, malformed syntax, or invalid N.
2. AST (items.rs): ItemData::Let.align: Option<u32> field.
3. IR (let_meta.rs): LetInfo.align: Option<u32> + with_align constructor.
4. cmd_build.rs: AST walk seeds LetInfo{mutable, align} from ItemData::Let.
   Routing block reads via let_meta().get().and_then(|i| i.align).
5. IR (data.rs): DataEntry.align widened u8 -> u32 (u8 capped at 255).
6. Emitter (writer.rs): add_rodata_bytes/add_data_bytes/add_bss_space
   signatures widened u8 -> u32. Internal `as u64` conversion unchanged.
7. Diagnostics (catalog.toml): P0250 (unknown), P0251 (malformed),
   P0252 (non-power-of-two or out-of-range).

## Alignment mechanism (MVP: shared .bss)

object crate's `Section::append_bss(size, align)` already promotes
`section.align = max(section.align, align)` and aligns the returned
offset. So passing align=4096 for pml4 lifts `.bss` to 4096-byte
alignment and places pml4 at the next 4096-aligned offset.

Per-symbol sections (.bss.pml4, .bss.pdpt) deferred to a follow-up
issue (PA10-006z) — MVP matches the current paideia-os link.ld model.

## Test coverage

- pa10_006y_align_pml4_bss: end-to-end fixture with `@align(4096)`
  on pml4; parses output ELF and asserts .bss sh_addralign >= 4096,
  pml4 in .bss, offset 4096-aligned.
- pa10_006y_align_default: regression guard — no `@align` keeps
  .bss alignment at 8.
- align_attr_errors: 4 diagnostic cases (P0250 unknown, P0251
  missing parens, P0252 non-power-of-two, P0252 zero).
- Parser unit test in parse_item.rs's mod tests.
- IR unit test in data.rs for DataEntry::new_bss with align=4096.
- SARIF insta snapshot regenerated for the 3 new diagnostic codes.

Full workspace green (only pre-existing paideia_os_r1_5_r2_5_rebuild
failure persists — unrelated R9/R10/R11 file-list staleness, filed
separately).

## Opportunistic side fix

The AST-walk pass in cmd_build.rs previously never wrote to
let_meta — every let_meta read in production returned None,
defaulting is_mutable to false. Now the walk seeds both `mutable`
and `align`, closing a latent gap. Data-side-table routing based
on `is_mutable` now works correctly for `let mut` bindings.

## Scope limitations

- Item-level `let` only (StmtData::Let untouched — MVP).
- Shared .bss section (no per-symbol sections).
- Only `align` recognized as attribute (@ident(args) reserved).
- Ceiling 2^30 (fits u32 + realistic MM alignments).

## paideia-os migration path (follow-up)

After submodule bump, boot_stub.S can migrate:

    let mut pml4 : [u64; 512] = uninit @align(4096)
    let mut pdpt : [u64; 512] = uninit @align(4096)
    let mut pd   : [u64; 512] = uninit @align(4096)

Currently reserved via .bss + link.ld ALIGN(4K) — additive.

+++ crates/paideia-as-emitter-elf/src/writer.rs
@@ -170,7 +170,7 @@ impl ElfWriter {
     /// Append `bytes` to the `.rodata` section with the specified alignment.
     /// Returns the offset at which the append starts.
     /// Phase-1 helper used for read-only data (constants, GDT descriptors, etc).
-    pub fn add_rodata_bytes(&mut self, bytes: &[u8], align: u8) -> u64 {
+    pub fn add_rodata_bytes(&mut self, bytes: &[u8], align: u32) -> u64 {
         let rodata_section = self.rodata_section_id();
         let offset = self
             .obj
@@ -186,7 +186,7 @@ impl ElfWriter {
     /// Append `bytes` to the `.data` section with the specified alignment.
     /// Returns the offset at which the append starts.
     /// Phase-1 helper used for initialized mutable data (Phase 6+).
-    pub fn add_data_bytes(&mut self, bytes: &[u8], align: u8) -> u64 {
+    pub fn add_data_bytes(&mut self, bytes: &[u8], align: u32) -> u64 {
         let data_section = self.obj.section_id(StandardSection::Data);
         let offset = self
             .obj
@@ -204,7 +204,7 @@ impl ElfWriter {
     /// Phase 6 m5-002: used for uninitialized mutable data (let mut x : T = uninit).
     /// Phase 6 m5-003: uses Section::append_bss() which doesn't write file data.
     /// The object crate correctly marks this as SHT_NOBITS and records the size.
-    pub fn add_bss_space(&mut self, size: u64, align: u8) -> u64 {
+    pub fn add_bss_space(&mut self, size: u64, align: u32) -> u64 {
         let bss_section = self.obj.section_id(StandardSection::UninitializedData);
         // Use the Section::append_bss() method which allocates space without writing to file.
         // This ensures .bss is properly marked as SHT_NOBITS with no file payload growth.
